Response messages
Added some response erorr messages for the sign-up form.
In this way, the user can spot the problem in the form and correct it.

Also, added a new `npm` package called `sweetalert2` in the
`package.json` file. This serves for the OK response message.

Signed-off by: Iulian Taiatu <iulian27_marius@yahoo.com>